# docs/generated/swagger/spec.json
{
  "produces": [
    "application/json"
  ],
  "schemes": [
    "http",
    "https"
  ],
  "swagger": "2.0",
  "info": {
    "description": "API for querying information about CockroachDB health, nodes, ranges,\nsessions, and other meta entities.",
    "title": "CockroachDB v2 API",
    "license": {
      "name": "Apache License, Version 2.0"
    },
    "version": "2.0.0"
  },
  "host": "localhost",
  "basePath": "/api/v2/",
  "paths": {
    "/databases/": {
      "get": {
        "description": "Lists all databases on this cluster.",
        "produces": [
          "application/json"
        ],
        "summary": "List databases",
        "operationId": "listDatabases",
        "parameters": [
          {
            "type": "integer",
            "description": "Maximum number of results to return in this call.",
            "name": "limit",
            "in": "query"
          },
          {
            "type": "integer",
            "description": "Continuation token for results after a past limited run.",
            "name": "offset",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Databases response",
            "schema": {
              "$ref": "#/definitions/databasesResponse"
            }
          }
        }
      }
    },
    "/databases/{database}/": {
      "get": {
        "description": "Returns the database's descriptor ID.",
        "produces": [
          "application/json"
        ],
        "summary": "Get database descriptor ID",
        "operationId": "databaseDetails",
        "parameters": [
          {
            "type": "string",
            "description": "Name of database being looked up.",
            "name": "database",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Database details response",
            "schema": {
              "$ref": "#/definitions/databaseDetailsResponse"
            }
          },
          "404": {
            "description": "Database not found"
          }
        }
      }
    },
    "/databases/{database}/grants/": {
      "get": {
        "description": "Returns grants on a database. Grants are the privileges granted to users\non this database.",
        "produces": [
          "application/json"
        ],
        "summary": "Lists grants on a database",
        "operationId": "databaseGrants",
        "parameters": [
          {
            "type": "string",
            "description": "Name of the database being looked up.",
            "name": "database",
            "in": "path",
            "required": true
          },
          {
            "type": "integer",
            "description": "Maximum number of grants to return in this call.",
            "name": "limit",
            "in": "query"
          },
          {
            "type": "integer",
            "description": "Continuation token for results after a past limited run.",
            "name": "offset",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Database grants response",
            "schema": {
              "$ref": "#/definitions/databaseGrantsResponse"
            }
          },
          "404": {
            "description": "Database not found"
          }
        }
      }
    },
    "/databases/{database}/tables/": {
      "get": {
        "description": "Lists names of all tables in the database. The names of all responses will\nbe schema-qualified.",
        "produces": [
          "application/json"
        ],
        "summary": "Lists tables on a database",
        "operationId": "databaseTables",
        "parameters": [
          {
            "type": "string",
            "description": "Name of the database being looked up.",
            "name": "database",
            "in": "path",
            "required": true
          },
          {
            "type": "integer",
            "description": "Maximum number of tables to return in this call.",
            "name": "limit",
            "in": "query"
          },
          {
            "type": "integer",
            "description": "Continuation token for results after a past limited run.",
            "name": "offset",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Database tables response",
            "schema": {
              "$ref": "#/definitions/databaseTablesResponse"
            }
          },
          "404": {
            "description": "Database not found"
          }
        }
      }
    },
    "/databases/{database}/tables/{table}/": {
      "get": {
        "description": "Returns details about a table.",
        "produces": [
          "application/json"
        ],
        "summary": "Get table details",
        "operationId": "tableDetails",
        "parameters": [
          {
            "type": "string",
            "description": "Name of the database being looked up.",
            "name": "database",
            "in": "path",
            "required": true
          },
          {
            "type": "string",
            "description": "Name of table being looked up. Table may be schema-qualified (schema.table) and each name component that contains sql unsafe characters such as . or uppercase letters must be surrounded in double quotes like \"naughty schema\".table.",
            "name": "table",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Database details response",
            "schema": {
              "$ref": "#/definitions/tableDetailsResponse"
            }
          },
          "404": {
            "description": "Database or table not found"
          }
        }
      }
    },
    "/events/": {
      "get": {
        "description": "Lists the latest event log entries, in descending order.",
        "produces": [
          "application/json"
        ],
        "summary": "List events",
        "operationId": "listEvents",
        "parameters": [
          {
            "type": "string",
            "description": "Type of events to filter for (e.g. \"create_table\"). Only one event type can be specified at a time.",
            "name": "type",
            "in": "query"
          },
          {
            "type": "integer",
            "description": "Filter for events with this targetID. Only one targetID can be specified at a time.",
            "name": "targetID",
            "in": "query"
          },
          {
            "type": "integer",
            "description": "Maximum number of results to return in this call.",
            "name": "limit",
            "in": "query"
          },
          {
            "type": "integer",
            "description": "Continuation token for results after a past limited run.",
            "name": "offset",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Events response",
            "schema": {
              "$ref": "#/definitions/eventsResponse"
            }
          }
        }
      }
    },
...
    "/users/": {
      "get": {
        "description": "List SQL users on this cluster.",
        "produces": [
          "application/json"
        ],
        "summary": "List users",
        "operationId": "listUsers",
        "parameters": [
          {
            "type": "integer",
            "description": "Maximum number of results to return in this call.",
            "name": "limit",
            "in": "query"
          },
          {
            "type": "integer",
            "description": "Continuation token for results after a past limited run.",
            "name": "offset",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Users response",
            "schema": {
              "$ref": "#/definitions/usersResponse"
            }
          }
        }
      }
    }
  },
  "definitions": {
...
    "DatabaseDetailsResponse_Grant": {
      "type": "object",
      "properties": {
        "privileges": {
          "description": "privileges are the abilities this grant gives to the user.",
          "type": "array",
          "items": {
            "type": "string"
          },
          "x-go-name": "Privileges"
        },
        "user": {
          "description": "user is the user that this grant applies to.",
          "type": "string",
          "x-go-name": "User"
        }
      },
      "x-go-package": "github.com/semistrict/ratel/pkg/server/serverpb"
    },
...
    "EventsResponse_Event": {
      "type": "object",
      "properties": {
        "event_type": {
          "description": "event_type is the type of the event (e.g. \"create_table\", \"drop_table\".",
          "type": "string",
          "x-go-name": "EventType"
        },
        "info": {
          "description": "info has more detailed information for the event. The contents vary\ndepending on the event.",
          "type": "string",
          "x-go-name": "Info"
        },
        "reporting_id": {
          "description": "reporting_id is the reporting ID for this event.",
          "type": "integer",
          "format": "int64",
          "x-go-name": "ReportingID"
        },
        "target_id": {
          "description": "target_id is the target for this event.",
          "type": "integer",
          "format": "int64",
          "x-go-name": "TargetID"
        },
        "timestamp": {
          "description": "timestamp is the time at which the event occurred.",
          "type": "string",
          "format": "date-time",
          "x-go-name": "Timestamp"
        },
        "unique_id": {
          "description": "unique_id is a unique identifier for this event.",
          "type": "array",
          "items": {
            "type": "integer",
            "format": "uint8"
          },
          "x-go-name": "UniqueID"
        }
      },
      "x-go-package": "github.com/semistrict/ratel/pkg/server/serverpb"
    },
...
    "TableDetailsResponse": {
      "description": "TableDetailsResponse contains grants, column names, and indexes for\na table.",
      "type": "object",
      "properties": {
        "columns": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/TableDetailsResponse_Column"
          },
          "x-go-name": "Columns"
        },
        "configure_zone_statement": {
          "description": "configure_zone_statement is the output of \"SHOW ZONE CONFIGURATION FOR TABLE\"\nfor this table. It is a SQL statement that would re-configure the table's current\nzone if executed.",
          "type": "string",
          "x-go-name": "ConfigureZoneStatement"
        },
        "create_table_statement": {
          "description": "create_table_statement is the output of \"SHOW CREATE\" for this table;\nit is a SQL statement that would re-create the table's current schema if\nexecuted.",
          "type": "string",
          "x-go-name": "CreateTableStatement"
        },
        "descriptor_id": {
          "description": "descriptor_id is an identifier used to uniquely identify this table.\nIt can be used to find events pertaining to this table by filtering on\nthe 'target_id' field of events.",
          "type": "integer",
          "format": "int64",
          "x-go-name": "DescriptorID"
        },
        "grants": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/TableDetailsResponse_Grant"
          },
          "x-go-name": "Grants"
        },
        "indexes": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/TableDetailsResponse_Index"
          },
          "x-go-name": "Indexes"
        },
        "range_count": {
          "description": "range_count is the size of the table in ranges. This provides a rough\nestimate of the storage requirements for the table.\nTODO(mrtracy): The TableStats method also returns a range_count field which\nis more accurate than this one; TableDetails calculates this number using\na potentially faster method that is subject to cache staleness. We should\nconsider removing or renaming this field to reflect that difference. See\nGitHub issue #5435 for more information.",
          "type": "integer",
          "format": "int64",
          "x-go-name": "RangeCount"
        },
        "stats_last_created_at": {
          "description": "stats_last_created_at is the time at which statistics were last created.",
          "type": "string",
          "format": "date-time",
          "x-go-name": "StatsLastCreatedAt"
        },
        "zone_config": {
          "$ref": "#/definitions/ZoneConfig"
        },
        "zone_config_level": {
          "$ref": "#/definitions/ZoneConfigurationLevel"
        }
      },
      "x-go-package": "github.com/semistrict/ratel/pkg/server/serverpb"
    },
    "TableDetailsResponse_Column": {
      "type": "object",
      "properties": {
        "default_value": {
          "description": "default_value is the default value of this column.",
          "type": "string",
          "x-go-name": "DefaultValue"
        },
        "generation_expression": {
          "description": "generation_expression is the generator expression if the column is computed.",
          "type": "string",
          "x-go-name": "GenerationExpression"
        },
        "hidden": {
          "description": "hidden is whether this column is hidden.",
          "type": "boolean",
          "x-go-name": "Hidden"
        },
        "name": {
          "description": "name is the name of the column.",
          "type": "string",
          "x-go-name": "Name"
        },
        "nullable": {
          "description": "nullable is whether this column can contain NULL.",
          "type": "boolean",
          "x-go-name": "Nullable"
        },
        "type": {
          "description": "type is the SQL type (INT, STRING, etc.) of this column.",
          "type": "string",
          "x-go-name": "Type"
        }
      },
      "x-go-package": "github.com/semistrict/ratel/pkg/server/serverpb"
    },
    "TableDetailsResponse_Grant": {
      "type": "object",
      "title": "Grant is an entry from SHOW GRANTS.",
      "properties": {
        "privileges": {
          "description": "privileges are the abilities this grant gives to the user.",
          "type": "array",
          "items": {
            "type": "string"
          },
          "x-go-name": "Privileges"
        },
        "user": {
          "description": "user is the user that this grant applies to.",
          "type": "string",
          "x-go-name": "User"
        }
      },
      "x-go-package": "github.com/semistrict/ratel/pkg/server/serverpb"
    },
    "TableDetailsResponse_Index": {
      "type": "object",
      "properties": {
        "column": {
          "description": "column is the column that this index indexes.",
          "type": "string",
          "x-go-name": "Column"
        },
        "direction": {
          "description": "direction is either \"ASC\" (ascending) or \"DESC\" (descending).",
          "type": "string",
          "x-go-name": "Direction"
        },
        "implicit": {
          "description": "implicit is an internal variable that's passed along.",
          "type": "boolean",
          "x-go-name": "Implicit"
        },
        "name": {
          "description": "name is the name of this index.",
          "type": "string",
          "x-go-name": "Name"
        },
        "seq": {
          "description": "seq is an internal variable that's passed along.",
          "type": "integer",
          "format": "int64",
          "x-go-name": "Seq"
        },
        "storing": {
          "description": "storing is an internal variable that's passed along.",
          "type": "boolean",
          "x-go-name": "Storing"
        },
        "unique": {
          "description": "unique is whether this a unique index (i.e. CREATE UNIQUE INDEX).",
          "type": "boolean",
          "x-go-name": "Unique"
        }
      },
      "x-go-package": "github.com/semistrict/ratel/pkg/server/serverpb"
    },
...
    "UsersResponse_User": {
      "type": "object",
      "title": "User is a CockroachDB user.",
      "properties": {
        "username": {
          "type": "string",
          "x-go-name": "Username"
        }
      },
      "x-go-package": "github.com/semistrict/ratel/pkg/server/serverpb"
    },
...
    "ZoneConfigurationLevel": {
      "description": "the object level at which the configuration is defined. This is needed\nbecause objects without a specifically indicated Zone Configuration will\ninherit the configuration of their \"parent\".",
      "type": "integer",
      "format": "int32",
      "title": "ZoneConfigurationLevel indicates, for objects with a Zone Configuration,",
      "x-go-package": "github.com/semistrict/ratel/pkg/server/serverpb"
    },
    "databaseDetailsResponse": {
      "type": "object",
      "title": "Response for databaseDetails.",
      "properties": {
        "descriptor_id": {
          "description": "DescriptorID is an identifier used to uniquely identify this database.\nIt can be used to find events pertaining to this database by filtering on\nthe 'target_id' field of events.",
          "type": "integer",
          "format": "int64",
          "x-go-name": "DescriptorID"
        }
      },
      "x-go-package": "github.com/semistrict/ratel/pkg/server"
    },
    "databaseGrantsResponse": {
      "type": "object",
      "title": "Response for databaseGrants.",
      "properties": {
        "grants": {
          "description": "Grants are the privileges granted to users on this database.",
          "type": "array",
          "items": {
            "$ref": "#/definitions/DatabaseDetailsResponse_Grant"
          },
          "x-go-name": "Grants"
        },
        "next": {
          "description": "The continuation token, for use in the next paginated call in the `offset`\nparameter.",
          "type": "integer",
          "format": "int64",
          "x-go-name": "Next"
        }
      },
      "x-go-package": "github.com/semistrict/ratel/pkg/server"
    },
    "databaseTablesResponse": {
      "type": "object",
      "title": "Response for databaseTables.",
      "properties": {
        "next": {
          "description": "The continuation token, for use in the next paginated call in the `offset`\nparameter.",
          "type": "integer",
          "format": "int64",
          "x-go-name": "Next"
        },
        "table_names": {
          "description": "TableNames contains the names of all tables in this database. Note that\nall responses will be schema-qualified (schema.table) and that every schema\nor table that contains a \"sql unsafe character\" such as uppercase letters\nor dots will be surrounded with double quotes, such as \"naughty schema\".table.",
          "type": "array",
          "items": {
            "type": "string"
          },
          "x-go-name": "TableNames"
        }
      },
      "x-go-package": "github.com/semistrict/ratel/pkg/server"
    },
    "databasesResponse": {
      "type": "object",
      "title": "Response for listDatabases.",
      "properties": {
        "databases": {
          "type": "array",
          "items": {
            "type": "string"
          },
          "x-go-name": "Databases"
        },
        "next": {
          "description": "The continuation token, for use in the next paginated call in the `offset`\nparameter.",
          "type": "integer",
          "format": "int64",
          "x-go-name": "Next"
        }
      },
      "x-go-package": "github.com/semistrict/ratel/pkg/server"
    },
    "eventsResponse": {
      "type": "object",
      "title": "Response for listEvents.",
      "properties": {
        "events": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/EventsResponse_Event"
          },
          "x-go-name": "Events"
        },
        "next": {
          "description": "The continuation token, for use in the next paginated call in the `offset`\nparameter.",
          "type": "integer",
          "format": "int64",
          "x-go-name": "Next"
        }
      },
      "x-go-package": "github.com/semistrict/ratel/pkg/server"
    },
...
    "tableDetailsResponse": {
      "title": "Response for tableDetails.",
      "$ref": "#/definitions/TableDetailsResponse"
    },
    "usersResponse": {
      "type": "object",
      "title": "Response for listUsers.",
      "properties": {
        "next": {
          "description": "The continuation token, for use in the next paginated call in the `offset`\nparameter.",
          "type": "integer",
          "format": "int64",
          "x-go-name": "Next"
        },
        "users": {
          "description": "usernames is a list of users for the CockroachDB cluster.",
          "type": "array",
          "items": {
            "$ref": "#/definitions/UsersResponse_User"
          },
          "x-go-name": "Users"
        }
      },
      "x-go-package": "github.com/semistrict/ratel/pkg/server"
    }
  },
  "securityDefinitions": {
    "api_session": {
      "description": "Handle to logged-in REST session. Use `/login/` to log in and get a session.",
      "type": "apiKey",
      "name": "X-Cockroach-API-Session",
      "in": "header"
    }
  }
}